requirements.txt
langgraph>=0.2.20,<0.3
langchain>=0.3.0,<0.4
langchain-core>=0.3.0,<0.4
langchain-community>=0.3.0,<0.4
langchain-groq>=0.2.0,<0.3
langchain-experimental>=0.3.0,<0.4
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
pydantic>=2.0.0,<3.0.0
tiktoken
pypdf
langchain_huggingface
